backend/requirements.txt
fastapi
uvicorn
websockets
python-multipart
python-jose[cryptography]
passlib[bcrypt]
sqlmodel
aiosqlite
authlib
httpx
itsdangerous
python-dotenv
